+++ src/data/whichever.rs
@@ -60,6 +60,27 @@
 /// When implementing [`Iterator`], the generics `F`, `B`, `R`, `P`, `S`, and
 /// `K` are reserved. Also, the `try_trait_v2` feature must be enabled.
 ///
+/// For [`Iterator`], it may also be desirable to have enum variants whose
+/// `Item` associated types differ, but use some map operation to unify them.
+/// This is possible with the following syntax:
+///
+/// ```
+/// #![feature(try_trait_v2)]
+/// use zoe::define_whichever;
+///
+/// define_whichever!{
+///     pub enum SignedOrUnsigned {
+///         Signed(std::iter::RepeatN<i32>),
+///         Unsigned(std::iter::RepeatN<u32>),
+///     }
+///
+///     #[map(i64::from)]
+///     impl Iterator for SignedOrUnsigned {
+///         type Item = i64;
+///     }
+/// }
+/// ```
+///
 /// ## Acknowledgements
 ///
 /// The syntax for matching generics was developed by
@@ -175,6 +196,7 @@ macro_rules! impl_traits {
 
     (@$dispatch:tt
         $(
+            $(#[map($map:expr)])?
             impl $(<
                 $( $lifetime:lifetime $(: $lifetime_bound:lifetime)? ),* $(,)?
                 $( $generics:ident
@@ -213,89 +235,89 @@ macro_rules! impl_traits {
             )?
             {
                 $($tt)*
-                $crate::impl_traits!{@$dispatch $trait}
+                $crate::impl_traits!{@[$dispatch $($map)?] $trait}
             }
         )*
     };
 
-    (@$dispatch:tt Read) => {
+    (@[$($dispatch:tt)*] Read) => {
         #[inline]
         fn read(&mut self, buf: &mut [u8]) -> ::std::io::Result<usize> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.read(buf))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.read(buf))
         }
 
         #[inline]
         fn read_vectored(&mut self, bufs: &mut [::std::io::IoSliceMut<'_>]) -> ::std::io::Result<usize> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.read_vectored(bufs))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.read_vectored(bufs))
         }
 
         #[inline]
         fn read_to_end(&mut self, buf: &mut ::std::vec::Vec<u8>) -> ::std::io::Result<usize> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.read_to_end(buf))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.read_to_end(buf))
         }
 
         #[inline]
         fn read_to_string(&mut self, buf: &mut ::std::string::String) -> ::std::io::Result<usize> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.read_to_string(buf))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.read_to_string(buf))
         }
 
         #[inline]
         fn read_exact(&mut self, buf: &mut [u8]) -> ::std::io::Result<()> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.read_exact(buf))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.read_exact(buf))
         }
     };
 
-    (@$dispatch:tt Write) => {
+    (@[$($dispatch:tt)*] Write) => {
         #[inline]
         fn write(&mut self, buf: &[u8]) -> ::std::io::Result<usize> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.write(buf))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.write(buf))
         }
 
         #[inline]
         fn flush(&mut self) -> ::std::io::Result<()> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.flush())
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.flush())
         }
 
         #[inline]
         fn write_vectored(&mut self, bufs: &[::std::io::IoSlice<'_>]) -> ::std::io::Result<usize> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.write_vectored(bufs))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.write_vectored(bufs))
         }
 
         #[inline]
         fn write_all(&mut self, buf: &[u8]) -> ::std::io::Result<()> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.write_all(buf))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.write_all(buf))
         }
 
         #[inline]
         fn write_fmt(&mut self, fmt: ::std::fmt::Arguments<'_>) -> ::std::io::Result<()> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.write_fmt(fmt))
+            $crate::impl_traits!(@[$($dispatch)*] self, [&mut], inner => inner.write_fmt(fmt))
         }
     };
 
-    (@$dispatch:tt Iterator) => {
+    (@[$dispatch:tt $($map:expr)?] Iterator) => {
         #[inline]
         fn next(&mut self) -> Option<Self::Item> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.next())
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner.next()$(.map($map))?)
         }
 
         #[inline]
         fn size_hint(&self) -> (usize, Option<usize>) {
-            $crate::impl_traits!(@$dispatch self, [&], inner => inner.size_hint())
+            $crate::impl_traits!(@[$dispatch] self, [&], inner => inner.size_hint())
         }
 
         #[inline]
         fn count(self) -> usize {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.count())
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner.count())
         }
 
         #[inline]
         fn last(self) -> Option<Self::Item> {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.last())
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner.last()$(.map($map))?)
         }
 
         #[inline]
         fn nth(&mut self, n: usize) -> Option<Self::Item> {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.nth(n))
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner.nth(n)$(.map($map))?)
         }
 
         #[inline]
@@ -304,7 +326,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(Self::Item),
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.for_each(f))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.for_each(f))
         }
 
         #[inline]
@@ -313,7 +335,7 @@ macro_rules! impl_traits {
             B: FromIterator<Self::Item>,
             Self: Sized,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.collect())
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.collect())
         }
 
         #[inline]
@@ -323,7 +345,7 @@ macro_rules! impl_traits {
             B: Default + Extend<Self::Item>,
             F: FnMut(&Self::Item) -> bool,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.partition(f))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.partition(f))
         }
 
         #[inline]
@@ -333,7 +355,7 @@ macro_rules! impl_traits {
             F: FnMut(B, Self::Item) -> R,
             R: ::std::ops::Try<Output = B>,
         {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.try_fold(init, f))
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner$(.map($map))?.try_fold(init, f))
         }
 
         #[inline]
@@ -343,7 +365,7 @@ macro_rules! impl_traits {
             F: FnMut(Self::Item) -> R,
             R: ::std::ops::Try<Output = ()>,
         {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.try_for_each(f))
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner$(.map($map))?.try_for_each(f))
         }
 
         #[inline]
@@ -352,7 +374,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(B, Self::Item) -> B,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.fold(init, f))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.fold(init, f))
         }
 
         #[inline]
@@ -361,7 +383,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(Self::Item, Self::Item) -> Self::Item,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.reduce(f))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.reduce(f))
         }
 
         #[inline]
@@ -370,7 +392,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(Self::Item) -> bool,
         {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.all(f))
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner$(.map($map))?.all(f))
         }
 
         #[inline]
@@ -379,7 +401,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(Self::Item) -> bool,
         {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.any(f))
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner$(.map($map))?.any(f))
         }
 
         #[inline]
@@ -388,7 +410,7 @@ macro_rules! impl_traits {
             Self: Sized,
             P: FnMut(&Self::Item) -> bool,
         {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.find(predicate))
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner$(.map($map))?.find(predicate))
         }
 
         #[inline]
@@ -397,7 +419,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(Self::Item) -> Option<B>,
         {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.find_map(f))
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner$(.map($map))?.find_map(f))
         }
 
         #[inline]
@@ -406,7 +428,7 @@ macro_rules! impl_traits {
             Self: Sized,
             P: FnMut(Self::Item) -> bool,
         {
-            $crate::impl_traits!(@$dispatch self, [&mut], inner => inner.position(predicate))
+            $crate::impl_traits!(@[$dispatch] self, [&mut], inner => inner$(.map($map))?.position(predicate))
         }
 
         #[inline]
@@ -416,7 +438,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(&Self::Item) -> B,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.max_by_key(f))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.max_by_key(f))
         }
 
         #[inline]
@@ -425,7 +447,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(&Self::Item, &Self::Item) -> ::std::cmp::Ordering,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.max_by(compare))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.max_by(compare))
         }
 
         #[inline]
@@ -435,7 +457,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(&Self::Item) -> B,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.min_by_key(f))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.min_by_key(f))
         }
 
         #[inline]
@@ -444,7 +466,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(&Self::Item, &Self::Item) -> ::std::cmp::Ordering,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.min_by(compare))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.min_by(compare))
         }
 
         #[inline]
@@ -453,7 +475,7 @@ macro_rules! impl_traits {
             Self: Sized,
             S: ::std::iter::Sum<Self::Item>,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.sum())
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.sum())
         }
 
         #[inline]
@@ -462,7 +484,7 @@ macro_rules! impl_traits {
             Self: Sized,
             S: ::std::iter::Product<Self::Item>,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.product())
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.product())
         }
 
         #[inline]
@@ -471,7 +493,7 @@ macro_rules! impl_traits {
             Self: Sized,
             F: FnMut(&Self::Item, &Self::Item) -> bool,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.is_sorted_by(compare))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.is_sorted_by(compare))
         }
 
         #[inline]
@@ -481,15 +503,15 @@ macro_rules! impl_traits {
             F: FnMut(Self::Item) -> K,
             K: PartialOrd,
         {
-            $crate::impl_traits!(@$dispatch self, [], inner => inner.is_sorted_by_key(f))
+            $crate::impl_traits!(@[$dispatch] self, [], inner => inner$(.map($map))?.is_sorted_by_key(f))
         }
     };
 
-    (@match_stmt $value:expr, [$($reference:tt)*], $pattern:pat => $result:expr) => {
+    (@[match_stmt] $value:expr, [$($reference:tt)*], $pattern:pat => $result:expr) => {
         match_macro!($value, $pattern => $result)
     };
 
-    (@wrapper_type $value:expr, [$($reference:tt)*], $pattern:pat => $result:expr) => {
+    (@[wrapper_type] $value:expr, [$($reference:tt)*], $pattern:pat => $result:expr) => {
         {
             let $pattern = $($reference)* $value.0;
             $result