+++ CHANGELOG.md
@@ -17,6 +17,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added accessor `byte_keys()` for `ByteIndexMap`
 - Added unchecked constructors for `ScalarProfile` and `StripedProfile`
 - Added `SamDataView` and `SamDataViewMut`
+- Added `map` attribute syntax to `define_whichever` and `impl_traits`, so that a map can be applied when implementing iterator
 - Added `score_from_path` and `viterbi` for `LocalPhmm` (behind `dev-phmm` feature gate)
 - Added `DomainPhmm`, along with `score_from_path` for it (behind `dev-phmm` feature gate)
 - Added `SemiLocalPhmm` (behind `dev-phmm` feature gate)